associate chats with rides, lost of mist formatting changes

+++ lib/services/supabase_service.dart
@@ -25,8 +25,7 @@ class SupabaseService {
     try {
       // TODO: set status to 'requested' once push notifications are added
       final authId = supabase.auth.currentUser!.id;
-      await supabase.from('follower').insert(
-          {'follower_user_id': authId, 'target_user_id': authIdToFollow, 'status': 'CONFIRMED'});
+      await supabase.from('follower').insert({'follower_user_id': authId, 'target_user_id': authIdToFollow, 'status': 'CONFIRMED'});
     } on PostgrestException catch (error) {
       //can also capture stacktrace with second argument after 'error'
       debugPrint(error.message);
@@ -36,11 +35,7 @@ class SupabaseService {
 
   static Future<void> unfollow(String authIdToUnfollow, String followerUserId) async {
     try {
-      await supabase
-          .from('follower')
-          .delete()
-          .eq('follower_user_id', followerUserId)
-          .eq('target_user_id', authIdToUnfollow);
+      await supabase.from('follower').delete().eq('follower_user_id', followerUserId).eq('target_user_id', authIdToUnfollow);
     } on PostgrestException catch (error) {
       debugPrint(error.message);
       rethrow;
@@ -50,19 +45,14 @@ class SupabaseService {
   static void removeFollower(String authIdToRemove) async {
     try {
       final authId = supabase.auth.currentUser!.id;
-      await supabase
-          .from('follower')
-          .delete()
-          .eq('follower_user_id', authIdToRemove)
-          .eq('target_user_id', authId);
+      await supabase.from('follower').delete().eq('follower_user_id', authIdToRemove).eq('target_user_id', authId);
     } on PostgrestException catch (error) {
       debugPrint(error.message);
       rethrow;
     }
   }
 
-  static void updatePassengerStatus(
-      String rideId, RidePassengerStatus newStatus, String passengerUserId) async {
+  static void updatePassengerStatus(String rideId, RidePassengerStatus newStatus, String passengerUserId) async {
     try {
       // create row in ride_passenger table
       //don't need to pass in intial status or created_at since those are created on the db side
@@ -79,8 +69,7 @@ class SupabaseService {
     }
   }
 
-  static Future<void> updateProfile(
-      String firstName, String lastName, String email, String phoneNumber, String bio) async {
+  static Future<void> updateProfile(String firstName, String lastName, String email, String phoneNumber, String bio) async {
     try {
       final authId = supabase.auth.currentUser!.id;
       final updates = {
@@ -116,6 +105,7 @@ class SupabaseService {
           .gte('datetime', DateTime.now())
           .not('driver_user_id', 'eq', user.id)
           .order('datetime', ascending: true);
+
       return result.map((item) => Ride.fromJson(item)).toList();
     } on PostgrestException catch (error) {
       debugPrint(error.message);
@@ -143,8 +133,7 @@ class SupabaseService {
           .from('ride_passenger')
           .select('passenger_user_id, status, profile(first_name, last_name)')
           .eq('ride_id', rideId);
-      List<PassengerProfile> ridePassengerProfile =
-          result.map((item) => PassengerProfile.fromJson(item)).toList();
+      List<PassengerProfile> ridePassengerProfile = result.map((item) => PassengerProfile.fromJson(item)).toList();
       return ridePassengerProfile;
     } on PostgrestException catch (error) {
       debugPrint(error.message);
@@ -172,15 +161,11 @@ class SupabaseService {
           .from('ride_passenger')
           .select('ride(*)')
           .inFilter('passenger_user_id', searchAuthIds)
-          .inFilter('status', [
-        RidePassengerStatus.confirmed.toShortString(),
-        RidePassengerStatus.requested.toShortString()
-      ]).lte('ride.datetime', DateTime.now());
+          .inFilter('status', [RidePassengerStatus.confirmed.toShortString(), RidePassengerStatus.requested.toShortString()]).lte(
+              'ride.datetime', DateTime.now());
 
-      List<Ride> result = passengerResults
-          .where((element) => element['ride'] != null)
-          .map((item) => Ride.fromJson(item['ride']))
-          .toList();
+      List<Ride> result =
+          passengerResults.where((element) => element['ride'] != null).map((item) => Ride.fromJson(item['ride'])).toList();
 
       final driverResults = await supabase
           .from('ride')
@@ -190,7 +175,7 @@ class SupabaseService {
           .order('datetime', ascending: false); //get activity data in descending datetime
 
       result.addAll(driverResults.map((item) => Ride.fromJson(item)));
-      result.sort((ride1, ride2) => ride1.dateTime.compareTo(ride2.dateTime)); //sort by date
+      result.sort((ride1, ride2) => ride2.dateTime.compareTo(ride1.dateTime)); //sort by date
       final noDupes = result.distinct((ride) => ride.id!);
       return noDupes.toList();
     } on PostgrestException catch (error) {
@@ -205,19 +190,11 @@ class SupabaseService {
           .from('ride_passenger')
           .select('ride(*)')
           .eq('passenger_user_id', authId)
-          .inFilter('status', [
-        RidePassengerStatus.confirmed.toShortString(),
-        RidePassengerStatus.requested.toShortString()
-      ]).lte('ride.datetime', DateTime.now());
-      List<Ride> result = passengerRides
-          .where((element) => element['ride'] != null)
-          .map((item) => Ride.fromJson(item['ride']))
-          .toList();
-      final driverRides = await supabase
-          .from('ride')
-          .select()
-          .eq('driver_user_id', authId)
-          .lte('datetime', DateTime.now());
+          .inFilter('status', [RidePassengerStatus.confirmed.toShortString(), RidePassengerStatus.requested.toShortString()]).lte(
+              'ride.datetime', DateTime.now());
+      List<Ride> result =
+          passengerRides.where((element) => element['ride'] != null).map((item) => Ride.fromJson(item['ride'])).toList();
+      final driverRides = await supabase.from('ride').select().eq('driver_user_id', authId).lte('datetime', DateTime.now());
       result.addAll(driverRides.map((item) => Ride.fromJson(item)));
       return result;
     } on PostgrestException catch (error) {
@@ -235,19 +212,12 @@ class SupabaseService {
           .from('ride_passenger')
           .select('ride(*)')
           .eq('passenger_user_id', supabase.auth.currentUser!.id)
-          .inFilter('status', [
-        RidePassengerStatus.confirmed.toShortString(),
-        RidePassengerStatus.requested.toShortString()
-      ]).gte('ride.datetime', DateTime.now());
-      List<Ride> result = passengerRides
-          .where((element) => element['ride'] != null)
-          .map((item) => Ride.fromJson(item['ride']))
-          .toList();
-      final driverRides = await supabase
-          .from('ride')
-          .select()
-          .eq('driver_user_id', supabase.auth.currentUser!.id)
-          .gte('datetime', DateTime.now());
+          .inFilter('status', [RidePassengerStatus.confirmed.toShortString(), RidePassengerStatus.requested.toShortString()]).gte(
+              'ride.datetime', DateTime.now());
+      List<Ride> result =
+          passengerRides.where((element) => element['ride'] != null).map((item) => Ride.fromJson(item['ride'])).toList();
+      final driverRides =
+          await supabase.from('ride').select().eq('driver_user_id', supabase.auth.currentUser!.id).gte('datetime', DateTime.now());
       result.addAll(driverRides.map((item) => Ride.fromJson(item)));
       result.sort((ride1, ride2) => ride1.dateTime.compareTo(ride2.dateTime));
       return result;
@@ -264,12 +234,9 @@ class SupabaseService {
           .select('profile!follower_target_user_id_fkey(*)')
           .eq('status', 'CONFIRMED')
           .eq('follower_user_id', authId);
-      List<Profile> friendsList = result
-          .where((element) => element['profile'] != null)
-          .map((item) => Profile.fromJson(item['profile']))
-          .toList();
-      friendsList.sort((prof1, prof2) =>
-          '${prof1.firstName}${prof1.lastName}'.compareTo('${prof2.firstName}${prof2.lastName}'));
+      List<Profile> friendsList =
+          result.where((element) => element['profile'] != null).map((item) => Profile.fromJson(item['profile'])).toList();
+      friendsList.sort((prof1, prof2) => '${prof1.firstName}${prof1.lastName}'.compareTo('${prof2.firstName}${prof2.lastName}'));
       return friendsList;
     } on PostgrestException catch (error) {
       debugPrint(error.message);
@@ -284,12 +251,9 @@ class SupabaseService {
           .select('profile!follower_follower_user_id_fkey(*)')
           .eq('status', 'CONFIRMED')
           .eq('target_user_id', authId);
-      List<Profile> friendsList = result
-          .where((element) => element['profile'] != null)
-          .map((item) => Profile.fromJson(item['profile']))
-          .toList();
-      friendsList.sort((prof1, prof2) =>
-          '${prof1.firstName}${prof1.lastName}'.compareTo('${prof2.firstName}${prof2.lastName}'));
+      List<Profile> friendsList =
+          result.where((element) => element['profile'] != null).map((item) => Profile.fromJson(item['profile'])).toList();
+      friendsList.sort((prof1, prof2) => '${prof1.firstName}${prof1.lastName}'.compareTo('${prof2.firstName}${prof2.lastName}'));
       return friendsList;
     } on PostgrestException catch (error) {
       debugPrint(error.message);
@@ -300,8 +264,7 @@ class SupabaseService {
   static Future<bool> isFollowing(String targetUserId, String currentUserId) async {
     try {
       List<Profile> followedProfiles = await getProfilesFollowed(currentUserId);
-      Profile targetProfileFound = followedProfiles
-          .firstWhere((element) => element.authId == targetUserId, orElse: () => Profile());
+      Profile targetProfileFound = followedProfiles.firstWhere((element) => element.authId == targetUserId, orElse: () => Profile());
       return targetProfileFound.authId == targetUserId;
     } on PostgrestException catch (error) {
       debugPrint(error.message);
